yolo/requirements.txt
opencv-python-headless
streamlit
ultralytics
numpy
PILLOW
torch
torchvision


# opencv-python
# opencv-python-headless

# # YOLOv5 requirements
# # Usage: pip install -r requirements.txt

# # Base ------------------------------------------------------------------------
# ultralytics
# opencv-python-headless
# #opencv-python
# streamlit==1.24.1
# gitpython>=3.1.30
# matplotlib>=3.3
# numpy>=1.18.5
# # opencv-python>=4.1.1
# Pillow>=7.1.2
# psutil  # system resources
# PyYAML>=5.3.1
# requests>=2.23.0
# scipy>=1.4.1
# thop>=0.1.1  # FLOPs computation
# torch
# # torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
# # torchvision>=0.8.1
# torchvision 
# tqdm>=4.64.0
# ultralytics>=8.0.111
# torch==1.12.1

# torch-geometric==2.1.0
# -f https://data.pyg.org/whl/torch-1.13.0+cpu.html
# torch-scatter
# -f https://data.pyg.org/whl/torch-1.12.1+cpu.html
# torch-sparse==0.6.16
# # protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# # Logging ---------------------------------------------------------------------
# # tensorboard>=2.4.1
# # clearml>=1.2.0
# # comet

# # Plotting --------------------------------------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0

# # Export ----------------------------------------------------------------------
# # coremltools>=6.0  # CoreML export
# # onnx>=1.10.0  # ONNX export
# # onnx-simplifier>=0.4.1  # ONNX simplifier
# # nvidia-pyindex  # TensorRT export
# # nvidia-tensorrt  # TensorRT export
# # scikit-learn<=1.1.2  # CoreML quantization
# # tensorflow>=2.4.0  # TF exports (-cpu, -aarch64, -macos)
# # tensorflowjs>=3.9.0  # TF.js export
# # openvino-dev  # OpenVINO export

# # Deploy ----------------------------------------------------------------------
# setuptools>=65.5.1 # Snyk vulnerability fix
# # tritonclient[all]~=2.24.0

# # Extras ----------------------------------------------------------------------
# # ipython  # interactive notebook
# # mss  # screenshots
# # albumentations>=1.0.3
# # pycocotools>=2.0.6  # COCO mAP
